requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0
google-cloud-texttospeech>=2.16.0
requests>=2.31.0
httpx>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
